Move memory related tests into dedicated file
