+++ Sources/ZIPFoundation/Archive.swift
@@ -106,13 +106,15 @@ public final class Archive: Sequence {
         static let size = 22
     }
 
+    private var preferredEncoding: String.Encoding?
+    private var memoryFile: MemoryFile?
     /// URL of an Archive's backing file.
     public let url: URL
     /// Access mode for an archive file.
     public let accessMode: AccessMode
+    public var data: Data? { return memoryFile?.data }
     var archiveFile: UnsafeMutablePointer<FILE>
     var endOfCentralDirectoryRecord: EndOfCentralDirectoryRecord
-    var preferredEncoding: String.Encoding?
 
     /// Initializes a new ZIP `Archive`.
     ///
@@ -175,7 +177,6 @@ public final class Archive: Sequence {
         setvbuf(self.archiveFile, nil, _IOFBF, Int(defaultPOSIXBufferSize))
     }
 
-    var memoryFile: MemoryFile?
     public init?(data: Data = Data(), accessMode mode: AccessMode, preferredEncoding: String.Encoding? = nil) {
         self.url = URL(string: "memory:")!
         self.accessMode = mode
@@ -213,8 +214,6 @@ public final class Archive: Sequence {
         }
         self.endOfCentralDirectoryRecord = endOfCentralDirectoryRecord
     }
-
-    public var data: Data? { return memoryFile?.data }
 
     deinit {
         fclose(self.archiveFile)
