+++ Tests/ZIPFoundationTests/ZIPFoundationWritingTests.swift
@@ -25,19 +25,6 @@ extension ZIPFoundationTests {
         XCTAssert(archive.checkIntegrity())
     }
 
-    func testCreateArchiveAddUncompressedEntryToMemory() {
-        let archive = self.memoryArchive(for: #function, mode: .create)
-        let assetURL = self.resourceURL(for: #function, pathExtension: "png")
-        do {
-            let relativePath = assetURL.lastPathComponent
-            let baseURL = assetURL.deletingLastPathComponent()
-            try archive.addEntry(with: relativePath, relativeTo: baseURL)
-        } catch {
-            XCTFail("Failed to add entry to uncompressed folder archive with error : \(error)")
-        }
-        XCTAssert(archive.checkIntegrity())
-    }
-
     func testCreateArchiveAddCompressedEntry() {
         let archive = self.archive(for: #function, mode: .create)
         let assetURL = self.resourceURL(for: #function, pathExtension: "png")
@@ -47,21 +34,6 @@ extension ZIPFoundationTests {
             try archive.addEntry(with: relativePath, relativeTo: baseURL, compressionMethod: .deflate)
         } catch {
             XCTFail("Failed to add compressed entry folder archive : \(error)")
-        }
-        let entry = archive[assetURL.lastPathComponent]
-        XCTAssertNotNil(entry)
-        XCTAssert(archive.checkIntegrity())
-    }
-
-    func testCreateArchiveAddCompressedEntryToMemory() {
-        let archive = self.memoryArchive(for: #function, mode: .create)
-        let assetURL = self.resourceURL(for: #function, pathExtension: "png")
-        do {
-            let relativePath = assetURL.lastPathComponent
-            let baseURL = assetURL.deletingLastPathComponent()
-            try archive.addEntry(with: relativePath, relativeTo: baseURL, compressionMethod: .deflate)
-        } catch {
-            XCTFail("Failed to add entry to compressed folder archive with error : \(error)")
         }
         let entry = archive[assetURL.lastPathComponent]
         XCTAssertNotNil(entry)
